[observer] Extract TaggedPatternClusterer into its own file
Move all tagging infrastructure (TagGroupByKey, TagGroupByKeyRegistry,
extractTagGroupByKey, globalClusterHash, TaggedPatternClusterer) from
log_pattern_extractor.go into log_tagged_pattern_clusterer.go.
No behaviour change.

Made-with: Cursor
